WIP Coordinate with email delivery service
The communication works fine between this service and the delivery
service. There is one exception : the function "setFormAsConverted". It
seems to work perfectly when tested directly though Virtuoso, but when
the real flow is tested some complaints form are converted 2, 3, .. 10
times. There is no apparent logic in the number of conversion. There is
always a moment where it stops, but it can be long.

To be investigated.